Address review comments WIP
