--- hazelcast/src/test/java/childfirstclassloader/TestProcessorMetaSupplier.java
@@ -1,64 +1,0 @@
-/*
- * Copyright (c) 2008-2021, Hazelcast, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package childfirstclassloader;
-
-import com.hazelcast.cluster.Address;
-import com.hazelcast.function.SupplierEx;
-import com.hazelcast.jet.core.ProcessorMetaSupplier;
-import com.hazelcast.jet.core.ProcessorSupplier;
-import org.example.jet.impl.deployment.ResourceCollector;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import java.util.List;
-import java.util.function.Function;
-
-public class TestProcessorMetaSupplier implements ProcessorMetaSupplier {
-
-    private SupplierEx<SupplierEx<String>> resourceReaderSupplier;
-    private SupplierEx<String> resourceReader;
-
-    public TestProcessorMetaSupplier(SupplierEx<SupplierEx<String>> resourceReaderSupplier) {
-        this.resourceReaderSupplier = resourceReaderSupplier;
-    }
-
-    @Override
-    public void init(@Nonnull Context context) throws Exception {
-        resourceReader = resourceReaderSupplier.get();
-        ResourceCollector.add("ProcessorMetaSupplier init " + resourceReader.get());
-    }
-
-    @Override
-    public Function<? super Address, ? extends ProcessorSupplier> get(List<Address> addresses) {
-        ResourceCollector.add("ProcessorMetaSupplier get " + resourceReader.get());
-        return address -> {
-            ResourceCollector.add("ProcessorMetaSupplier create " + resourceReader.get());
-            return new TestProcessorSupplier(resourceReaderSupplier);
-        };
-    }
-
-    @Override
-    public void close(@Nullable Throwable error) throws Exception {
-        ResourceCollector.add("ProcessorMetaSupplier close " + resourceReader.get());
-    }
-
-    public static TestProcessorMetaSupplier create() {
-        return new TestProcessorMetaSupplier(
-                () -> new ResourceReader()
-        );
-    }
-}
