+++ hazelcast/src/test/java/com/hazelcast/jet/impl/deployment/ChildFirstClassLoaderTest.java
@@ -57,6 +57,8 @@ public class ChildFirstClassLoaderTest {
 
     @Test
     public void urlsMustNotBeNullNorEmpty() {
+        assertThatThrownBy(() -> new ChildFirstClassLoader(null, ClassLoader.getSystemClassLoader()))
+                .isInstanceOf(NullPointerException.class);
         assertThatThrownBy(() -> new ChildFirstClassLoader(new URL[0], ClassLoader.getSystemClassLoader()))
                 .isInstanceOf(IllegalArgumentException.class);
     }
